requirements.txt
discord-py-interactions==3.0.2
discord.py==1.7.3
requests==2.25.0
tweepy==4.4.0
